add more robust error handling when writing info.db for tables
server ran out of disk space, again - found that info.db and info.db.bak
were both empty

also, fix broken aggregation test

+++ src/lib/aggregate_test.go
@@ -171,7 +171,7 @@ func TestHistograms(test *testing.T) {
 			val := ages[index]
 
 			// TODO: margin of error should be less than 1!
-			if math.Abs(float64(v-val)) > 1 {
+			if math.Abs(float64(v-int64(val))) > 1 {
 				test.Error("P", k, "VAL", v, "EXPECTED", val)
 			}
 		}
